mymini/requirements.txt
alembic==1.8.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==1.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.0
Flask-WTF==1.0.1
greenlet==1.1.3.post0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
Pillow==9.2.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
WTForms==3.0.1
